Restore comment forms automatically, not on demand
It is risky and inconvenient for the user to fork the behavior into automatic comment forms restoring (when the user wasn't warned on leaving) and on-demand (when the user was warned on leaving). The user may leave because they want to go to a different page or make the page not produce notifications, not because they don't want to submit the forms. It is also easy to miss the link restoring the forms and close the notification instead.