Don't use getters and private class fields
Using private class fields (this.#property) hampers debugging and is not consistent with private methods that have even less support and were not used. Using getters is perhaps not the best practice because that introduces implicit manipulations that can be avoided. Comment#$elements and Section#$elements getters and setters are kept so far.