Use gendered namespace for user links autocomplete in some cases
Namely, cases when the gender is available (for example, if we're replying to a user and their gender was already loaded to form the edit summary correctly). It's too costly to request the gender preliminarily for every user in a menu, or to wait until the request is fulfilled when the user selects a user from the menu, so we have to create a configuration property for this, as there is currently no way to get that value from API, see https://phabricator.wikimedia.org/T204610.