Change way autocomplete values are inserted
Now by default they are inserted in the simple form, if Shift is held--in the advanced form allowing to type in custom values (for example, [[Article|<text>]], if Ctrl is held--mention is inserted with colon separator after the user link.

The selection offsets handling is incorporated into Tribute code. The format of autocomplete collection configs has also changed.