Remove unseenCount variable from navPanel module
Calculate the number of unseen comments directly every time instead. We needed the variable mostly because of very frequent uses in eventHandlers.registerSeenComments (navPanel.registerSeenComments at the time). Now that it is gone from there, we no longer need it.

Also remove navPanel.decrementUnseenCount and navPanel.getUnseenCount methods, as they had meaning only as gateways to the unseenCount variable.

Fix not checking for the Comment#isFlashOnSightSet property in Comment#registerSeen in process.