Modify English text: sd-removedata-help (#19)
The message could be a bit misleading: "If X, you will have to remove the data from there too" --> Does it mean that I *have to* remove the data from other domains if I want to remove it from the current domain? (what consequences would I face if I don't remove it in other domains? would the data in other domains be corrupted?)

Also, changed "domain" to "wiki", as it was too technical.

I'm not a native English speaker, so feel free to correct me or submit you own patch if the text if it can be improved.