Streamline processPage() based on several categories of page activeness
In the process, fix losing session when an old revision page is closed and empty session saved. Fix processing the "editondblclick" and "editsectiononrightclick" options (they didn't make any difference as the module they rely on in the code no longer exists). Don't suppress the behaviors brought by these options, as it was before.