Don't insert heading markup if headline is not entered
This is done mostly for the ability to use substituted templates containing a heading in them. The other possible use case is when the "nosummary" parameter is missed in the preload configuration (it works to disable the headline input in this case), but the headline input is left alone by the user (as was here before the fix: https://ru.wikipedia.org/w/index.php?diff=112451882).