Apply many minor improvements
Rename global object properties starting with "CURRENT_", removing "CURRENT_" (for example, cd.g.CURRENT_PAGE → cd.g.PAGE). Rename a couple of variables and properties starting with "this", replacing "this" with "current", when they are unrelated to the "this" object (for example, thisPageVisits → currentPageVisits).

Use util.addToArrayIfAbsent and util.removeFromArrayIfPresent to add an element to the array if it's not already present and remove an element from the array if it's present.