Don't change URL hash in part of cases
Namely, when using page navigation (pageNav.js) and comments' up and down buttons to navigate sections and comments. The benefit of changing the URL is little (being able to use the "Back" button to go back), but the cost is big (cluttering the browser history).

This commit partly reverts 78a134207dbbe13c06424190f7d4dcd3a3632237.