Don't insert replies before "outdent" templates
Currently if an "outdent" template is in the middle of the thread that should be above the place of insertion of a reply, the reply is placed after the target reply (which could not be the best possible solution).

Improve boot~initPatterns, modify util.caseInsensitiveFirstCharPattern to replace spaces as well.