Improve algorithm and text for "Comment not found" messages
Improve searching for comments in the archive. Show "The previous comment from this user by time" even if we found some comment in the archive (that could be not correct since we search by 4 timestamps without an author). Use a 3 minute gap instead of a 5 minute gap, as was done in Comment.getByAnchor.

Change-Id: I29bfffafaf30297b9643d03899e973c0c934fe69