Use standard error messages from API in HTML format in many occasions
Specify errorformat: 'html' and related parameters in API requests instead of parsing raw wikitext from the response in the legacy format. Try to automatically detect framed messages to show them as they are.

Keep the i18n strings so far in case short error messages are required somewhere (for instance, in the move section dialog).

Change-Id: Iddcb72032e7af71836815873f1a5102f830d36c4