Remove some of DT markup before parsing pages
According to T273072, a change will be deployed starting today that will add the DiscussionTools markup to pages even if DT is disabled for the user. CD already parses comment links from notifications (which seems to be this markup's purpose for disabled DT). Unless the elements prove useful to CD or other scripts, it's better to get rid of them rather than deal with them one by one while parsing.

Use user options to tell the conflict with DT instead of looking for Reply Tool's markup.

Change-Id: I87cf44eb4d7b65e2eb7f2908b5de4cadfbc67e8c