Replace "Add topic" tab link to CD's on most pages
Initially, this behavior was only on 404 pages in edit mode, but it makes sense to spread this behavior to all pages that are eligible to run CD where "wgAction" is not "view" (there, the standard link stays as a fallback way to add a topic). Also always replace "Add topic" link if DT's New Topic Tool is enabled, because the standard editing interface is never opened in that case for adding a topic.

Change-Id: I51e4e88f76656f6b4630df3c244dda5a1c2685f9