Make CD compatible with DT's new topic tool
The main purpose of this change is to harness DT's rethinking of &action=edit&section=new page as an article page (wgIsArticle=true) where an "Add topic" form is placed. On such pages, we replace DT's form with ours, keeping content that was already typed.

Add "Add topic" form as a child of #mw-content-text, outside of .mw-parser-output, to ensure it is at the bottom of the page.

Change-Id: Icb9b63867fedaf7851908525c7975a4989abc7f8