Refactor page targets extraction
The impetus for this refactoring is the need to know the current section while parsing a comment. This allows to avoid having a comment that stretches between sections for the reason that the section heading is wrapped in a tag (example: https://meta.wikimedia.org/wiki/Community_Wishlist_Survey_2015/Editing/chy) and is not immediately visible while parsing the comment.

SectionSkeleton#elements as well as Section#$elements are discarded - we limit ourselves to the lastElement property which, together with Section#$heading, covers most of the use cases.

Change-Id: Iaa2d9b3288d270a273259321022ba32ac8f09831