Improve how comments look when copied
When comments reformatting is on, the comment header and footer were copied entirely (with no space delimiting the author and date), while there is only a need to copy the author and date, as a rule. This commit fixes it, wrapping author links in a wrapper and setting "user-select: none" style for irrelevant elements.

Change-Id: Ifcc380f5d87675a654aef60456eb6ec53cb83cad