Add and improve quoting and pasting functionality
If the user presses "Q" or replies (the button gets renamed to "Quote") to a comment while a comment is selected, the opened comment form receives the selection.

Paste events are now handled, and wikitext may be pasted instead of plain text after a REST API call. The conversion is also performed when quoting.

cd.config.quoteFormatting now can be an object, containing a property for multiline quotes.

Add cd.g.ALLOWED_TAGS. Extract eventHandlers.handleMouseMove().

Change-Id: Iddb479358da86f86e658e5c10e2744c5361bdcf1