Remove redundant error checks in apiWrappers module
Checks for resp.error seem to be completely irrelevant - this kind of error is thrown by mw.Api. Other checks for property existence are perhaps redundant - they are very unlikely, and if they occur, there is no difference between a CD error and a native JavaScript error that would be thrown. See "Let it crash".

Change-Id: Id61b24fbc31b0bcb4575e636b03d8bd924331a69