Fix JS error in updateChecker~mapComments
Extract updateChecker~sortCommentsByMatchScore. Fix the calculation of the partsMatchedProportion variable - previously is dependent on which comment is a target and which is a candidate. Now it will have the same value in both situations. This problem has led to sortCommentsByMatchScore() returning no array elements despite earlier it returned at least one element with the same comments supplied (but in different parameters). This should be a result of some comment falling under 1.66 score threshold.

Change-Id: Id2eb4aea495acb1555f0ebe44d3bed9a0ce23081