Redesign section menu, show metadata under section headings
Use an under-the-heading menu instead of expanding the existent [ section menu ]. The new narrower Vector poses a problem for the current design, as headlines are too often stretched to the full width of the content area, making the menu to clumsily wrap to the next line.

Also add a line with metadata (the number of comments, including new ones, the number of authors, and a linked date of the last comment) under every 2-level section heading.

This change introduces a design similar to one that will be introduced to DiscussionTools by T310914.

Extract Section#canAddReply, Section#canAddSubsection, and Section#showAddSubsectionButtonOnReplyButtonHover methods.

Change-Id: I3fd1425bc25aa1cd44e0a7164b0a8d2a13f14d9a