Change how highlighting of target comments works
Now the highlighting color is blue, unless it's a comment just posted. Multiple comments can be passed to the boot process to highlight them all. Now Comment#scrollTo takes an object instead of a list of parameters. Extract BootProcess#processTarget from BootProcess#processFragment. Conflict with autocollapse is resolved in BootProcess#processTarget with the help of Comment#expandAllThreadsDownTo, not in Thread~autocollapseThreads. Fix resetting the fragment on page reload.

Change-Id: Ia87b67fbbd3905875651048f6fd1c2c406d816ff