Fix processing outdent templates
Comments at 0th and 1st levels were incorrectly assumed to be outdented together with a comment directly outdented if they were at the same level as the outdented comment and followed it.

Also, set isOutdented to true on all comments outdented together with the directly outdented comment.

Extract CommentSkeleton#updateOutdentWidth.

Change-Id: Id9e3d3377b7479004a6a7a27f32021397e66190d