Implement automatic outdent functionality
Add the relevant setting.

Extract Comment#findProperPlaceForReply, Comment#matchProperPlaceRegexps, Comment~getAdjustedChunkCodeAfter from Comment#modifyWholeCode. Update modifyWholeCode method of various classes to return the comment code as well.

Link to page editing instead of section editing when a comment or proper place were not found in the code. Also, don't forget the comment form (in case the user forgot to copy the code). This was done previously when I assumed the error will always appear in the beginning of the editing, when the user didn't type much. But that is not guaranteed.

Use logical assignment operators.

Change-Id: I683a098e6ddcf31b9eb9cb19af7fd80045c6cfbe