Split instance and static properties of some classes
Instance and static properties of Comment, Section and CommentForm classes are kept in separate files, but static properties are merged into the main object. This change abolishes the merge so that VS Code and other IDEs could pick up where to look for static methods.

The objects with static properties are now called CommentStatic, SectionStatic and CommentFormStatic. In fact, these are now plain objects, but I kept the capitalization to keep the visual resemblance to their class counterparts.

Change-Id: I56d707308ecd8f5308a2fb56a68367e56a0c844b