i18n: Change how new comment count messages are formed and used
"N comments +M new" might be ambiguous (is the total comment count N+M? - no), but In some languages (such as Hebrew) it is unclear, as was reported. To allow for another composition of the output (for instance, using a comma before the new comment count), we merge the original comment count and new comment count messages. (But the comment count message stays - it is now shown when there are no new messages).

Change-Id: I46249717bef6e1fe9c1d6cce4499196c6bb2d168