Unhide all sections when page search box is opened
This alleviates the problem that you can't search the full page when the "Improve performance" setting is enabled. A catch is that we can't say why a page lost focus its focus. We can only detect whether it became invisible after that (after half a second or so). So we try not to hide all sections when the page becomes invisible completely, because unhiding them could take time and be noticeable to the user when they refocus the page.

Change-Id: Id261861fe67bb010dd91ce6251baa2b33e22cfdb