Unhide all sections on window blur even if page becomes invisible
This applies to users who have "Try to improve performance" setting enabled.

If we wait for 500ms as we did, the search box would show the incorrect number of results both in Chrome and Firefox, which would mislead the user (even though they can hit Enter and get the counter updated). Here we reconcile with the fact that unhiding sections can take time and be noticeable to the user when they refocus the page.

Change-Id: I152bc909678c58cb61ba1adb1b50557b38ecac03