Support new heading markup on a more fundamental level
The support of the new heading markup (see T314714) in fcca373a76c130c7fc144270644803fd01bff8c7 was superficial, and it was broken in some places.

Here, we introduce `isWrapper` and `level` properties for headings and `hElement` property for sections, supported for both old and new markup, both in the window and web
worker context. (How classes are handled in the worker context is reworked in the process; the `className` property is now supported.)

Change-Id: Ia419737c3d09b52999572f8c33665108b1a9da3a