Make DT comment links into span elements
It doesn't make sense to retain links in the old design as 1) there is the "link" menu item implementing that functionality, 2) the overlay comment menu often obstructs the timestamp.

Using a span element instead of outright removing it (while keeping its contents) allows to retain the styling, like the gray color.

Change-Id: Ic67e6bf3efe1ff794585fa9a24d8a3f4d8bd986e