Simplify logic for updating thread lines
`mutate` events are not triggered when the page is being reloaded, so the condition is mostly redundant. Even if Thread#updateLines() is called while the page is being reloaded, commentRegistry will be empty in cases that matter.

This fixes the issue where lines were not updated after autocollapsing threads. Lines are also now updated on the visibilitychange event in case mutate events are not triggered for some reason after the document becomes visible.

Change-Id: Ib6074154d8d262af735c4db7487507e37bdbb292