Don't recalculate positions of all comments every second
Halting the recalculations at the 3rd comment with an unchanged position behavior broke in d7d5687a4ec908fd3023adce3a99b5d6ac396362. And the fact it was broken means that the low value of notMovedCount was not the culprit of misplaced comment backgrounds reported by users that I thought it was. Also, since it was broken and everything seemingly worked fine, means that I gotta check how big the impact of recalculating comment positions is.

Change-Id: Ia98e8f5c6ea49fc9f4af7cd7345f0e7baabada9a