Change behavior of subscribe buttons in archives
Per my comment at T365320#9811020, having subscribe buttons for topics may actually be the correct behavior, since sometimes topics do get unarchived. And if a topic I'm subscribed to gets archived, it makes sense for me to be able to unsubscribe from it. As for pages subscription, the subscribe button on them is likely useless.

Change-Id: I77a6349302cb5f8729bbb0abe8e40530f46024b5