Improve comment marker and thread line positioning
With 140% scale the hovered thread line was misaligned in some cases with the marker in Chrome. I despaired of trying to make sense of how Chrome handles this and decided that it would be better to resort to using equal values for devicePixelRatioToDivisor, even though this doesn't address the root problem.

Separately, fix the gap closer width.

Change-Id: Id09898e9a029d7545910bddb5d1125f7e796b842