Fix submitting "Add section" form
The "Add topic" button was not restored. The subscription button was not added (broken in 20bf0954c13319d2127952403bf06e7bc34acc3f when mw.config.set(parseData.jsconfigvars); was moved). Tear down the comment form when rescuing it as well (so it's not torn down only when submitting anything other than an "Add section" form).

Change-Id: I3ce9ac819b327fd6522dc5c367c697244453ffbc