Simplify thread navigation with middle mouse held
When scrolling down as the first move, don't scroll to the last comment in the thread unless it's the last thread available. Always scroll to the next thread. Also, take 15 pixels of clearance into account when calculating deltas (it was only taken into account on first mouse moves, not on further ones). When navigating a thread, don't align collapsed threads to the bottom of the viewport when the thread expand note is visible.

Change-Id: I79cfa87601dc82567acb98fe44b3a7d95d27311a